Cope with Google accounts that haven't been provisioned yet
When a user is assigned a netid@nyu.edu email address, nothing is
created on the Google side until they first log in.  This causes
problems when a Sakai group is synced to Google and one of these
not-yet-provisioned addresses is sent over.

What you might think would happen: a placeholder membership is added
to the group.  When the user's account is provisioned, those
placeholders are resolved and linked to the real user account.
The user then sees their group membership.

What actually happens: the placeholder has no effect.  The group will
report that address as a member, but the user can't access it.

We want to avoid this issue by only sending email addresses that we
know have already been provisioned on the Google side.  To do this, we
hit the User API and search for an email address prior to adding it.
We can only do this for accounts on the Google domain; other addresses
are assumed to exist already.

So our strategy is:

  * When a new email shows up in a group, mark it as needing a lookup

  * Whenever our main loop runs, we check accounts that haven't
    been marked off as provisioned.  That covers users we've never
    checked, and users we've checked but not found.  Since some users
    might never turn up, the longer an account has been marked as "not
    provisioned", the less frequently we check it.

  * When a user becomes provisioned, we resync all of the groups
    they're in.