Handle deletions before anything else
When a user's email address changes we send a delete for the old one
and an insert for the new one.  Our code doesn't know whether those
email addresses are aliases for the same account, but Google does.
We'll send the delete first to try to ensure that we leave the user in
the group when this happens.

The next run would have set things right anyway, but this might help
to avoid brief periods of unavailability for those users.